fix(seo): wire fallback timeline splitter into static job pages
Every crawled job ships without `_canonical` (no AI pipeline produces it;
`translate-pending.yml` only relocalises text), so the no-canonical fallback
path in `build-plugins/jobsSeoPagesPlugin.ts` is taken for 100% of the
~6,144 jobs in `data/jobs.json`. That path was emitting only the
"Panoramica" section + an empty `<div class="timeline">`, while the
hydrated `JobBoard` component split the same raw description into 6
sections (Mansioni, Competenze, Cosa offre, Contatti, Azienda, Dettagli).
Result: crawlers and curl-based scrapers saw a stub body; users saw the
full content.

- Extract the heuristic splitter (~700 LoC) from
  `components/community/JobBoard.tsx` into a pure module at
  `services/jobs/canonicalFallback.ts`. No behaviour change for the SPA
  (existing JobBoard fallback test still passes byte-for-byte; the
  function is re-exported from JobBoard so external test imports still
  work).
- Wire `buildFallbackCanonicalContent` into the build plugin: when
  `readCanonicalByLocale` returns null or an empty payload (every job
  today), synthesize a `CanonicalLocaleContent` from the raw description
  + requirements. Downstream `canonicalSummary` / `canonicalSections` /
  `canonicalResponsibilities` / etc. now have content to render so the
  static HTML emits 4-6 `<div class="timeline-step">` blocks matching
  the SPA.
- Regression test `tests/seo/canonical-fallback-shape.test.ts` pins the
  splitter against a real Tether description (the one originally
  flagged) and asserts >=3 timeline buckets populate.

Verified locally: `buildFallbackCanonicalContent` on the Tether job
yields summary(3), responsibilities(3), requirements(2), benefits(10),
contacts(12), company(1), details(1) — matches the SPA output the
crawler-facing HTML was missing.
